+++ src/components/Project/Resource/ResourceFilterBar.tsx
@@ -19,6 +19,9 @@ interface ResourceFilterBarProps {
     onViewChangeClick?: () => void;
     isTableView?: boolean;
     icon: React.ReactElement<SvgIconProps>;
+    selectedItemsCount?: number;
+    resetSelectedItemsCount?: () => void;
+    onBatchDeleteClick?: () => void;
 }
 
 const ResourceFilterBar: React.FC<ResourceFilterBarProps> = ({
@@ -32,7 +35,9 @@ const ResourceFilterBar: React.FC<ResourceFilterBarProps> = ({
     onCreateClick,
     additionalCreateClick,
     createLabel = "Create Item", // Default label for the create button
-    addtionalCreateLabel = "Additional Create Item"
+    addtionalCreateLabel = "Additional Create Item",
+    selectedItemsCount,
+    onBatchDeleteClick
 }) => {
     return (
         <Box display="flex" justifyContent="space-between" alignItems="center" mb={2}>
@@ -73,6 +78,11 @@ const ResourceFilterBar: React.FC<ResourceFilterBarProps> = ({
                         sx={{ ml: 1 }}
                     >
                         {addtionalCreateLabel}
+                    </Button>
+                )}
+                {Boolean(selectedItemsCount) && (
+                    <Button variant="soft" color="danger" sx={{ ml: 1 }} onClick={onBatchDeleteClick} size="sm">
+                        Delete Selected ({selectedItemsCount})
                     </Button>
                 )}
             </Box>