+++ src/components/Project/Thumbnails/TableThumbnail.tsx
@@ -5,13 +5,10 @@ export const TableThumbnail = () => {
     return (
         <div
             style={{
-                position: "relative",
-                width: "100%",
                 height: 160,
                 display: "flex",
                 justifyContent: "center",
-                alignItems: "center",
-                padding: 20
+                alignItems: "center"
             }}
         >
             <img src={TableIcon} alt="table" style={{ objectFit: "contain", maxWidth: "100%", maxHeight: "100%" }} />