+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Batch select and delete resources [#140](https://github.com/IN-CORE/incore-studio/issues/140)
+
 ## [Alpha-4] - 04-11-2025
 
 ### Fixed